collision and map functions. Added tile-based map. changed collisions to use tile based maps.

+++ game_1/sprites.h
@@ -8,17 +8,17 @@ extern const uint8_t CharacterWALK2[32][32];
 extern const uint8_t CharacterDASH[32][32];
 extern const uint8_t CharacterJUMP[32][32];
 
-extern const uint32_t cat_walk1[32][32];
-extern const uint32_t cat_walk2[32][32];
-extern const uint32_t cat_walk3[32][32];
-extern const uint32_t cat_walk4[32][32];
+extern const uint8_t cat_walk1[32][32];
+extern const uint8_t cat_walk2[32][32];
+extern const uint8_t cat_walk3[32][32];
+extern const uint8_t cat_walk4[32][32];
 
 // mouse sprites
-extern const uint32_t mouse_sat[16][16];
-extern const uint32_t mouse_walk1[16][16];
-extern const uint32_t mouse_walk2[16][16];
+extern const uint8_t mouse_sat[16][16];
+extern const uint8_t mouse_walk1[16][16];
+extern const uint8_t mouse_walk2[16][16];
 
 // environment sprites
-extern const uint32_t grass_block[16][16];
+extern const uint8_t grass_block[16][16];
 
 #endif // SPRITES_H
